+++ Interface/AddOns/NDui/Modules/Misc/QuickJoin.lua
@@ -7,22 +7,19 @@ local TT = B:GetModule("Tooltip")
 	QuickJoin 优化系统自带的预创建功能
 	1.双击搜索结果，快速申请
 	2.自动隐藏部分窗口
-	3.美化LFG的相关职业图标
-	4.自动邀请申请
-	5.显示队长分数，并简写集市钥石
+	3.自动邀请申请
+	4.显示队长分数，并简写集市钥石
 ]]
-local select, wipe, sort, gsub, tremove = select, wipe, sort, gsub, tremove
+local select, gsub, tremove = select, gsub, tremove
 local StaticPopup_Hide, HideUIPanel, GetTime = StaticPopup_Hide, HideUIPanel, GetTime
-local UnitIsGroupLeader, UnitClass, UnitGroupRolesAssigned = UnitIsGroupLeader, UnitClass, UnitGroupRolesAssigned
+local UnitIsGroupLeader = UnitIsGroupLeader
 local C_Timer_After, IsAltKeyDown = C_Timer.After, IsAltKeyDown
 local C_LFGList_GetSearchResultInfo = C_LFGList.GetSearchResultInfo
 local C_LFGList_GetActivityInfoTable = C_LFGList.GetActivityInfoTable
-local C_LFGList_GetSearchResultMemberInfo = C_LFGList.GetSearchResultMemberInfo
 local C_ChallengeMode_GetMapUIInfo = C_ChallengeMode.GetMapUIInfo
 
 local HEADER_COLON = _G.HEADER_COLON
 local LE_PARTY_CATEGORY_HOME = _G.LE_PARTY_CATEGORY_HOME or 1
-local LFG_LIST_GROUP_DATA_ATLASES = _G.LFG_LIST_GROUP_DATA_ATLASES
 local scoreFormat = DB.GreyColor.."(%s) |r%s"
 
 local LFGListFrame = _G.LFGListFrame
@@ -54,162 +51,6 @@ end
 function M:HookDialogOnShow()
 	pendingFrame = self
 	C_Timer_After(1, M.DialogHideInSecond)
-end
-
-local roleCache = {}
-local roleOrder = {
-	["TANK"] = 1,
-	["HEALER"] = 2,
-	["DAMAGER"] = 3,
-}
-local indexToRole = {"TANK", "HEALER", "DAMAGER"}
-
-local function sortRoleOrder(a, b)
-	if a and b then
-		return a[1] < b[1]
-	end
-end
-
-local function GetPartyMemberInfo(index)
-	local unit = "player"
-	if index > 1 then unit = "party"..(index-1) end
-
-	local class = select(2, UnitClass(unit))
-	if not class then return end
-	local role = UnitGroupRolesAssigned(unit)
-	if role == "NONE" then role = "DAMAGER" end
-	return role, class, UnitIsGroupLeader(unit)
-end
-
-local function GetCorrectRoleInfo(frame, i)
-	if frame.resultID then
-		local role, class = C_LFGList_GetSearchResultMemberInfo(frame.resultID, i)
-		return role, class, i == 1
-	elseif frame == ApplicationViewerFrame then
-		return GetPartyMemberInfo(i)
-	end
-end
-
-local function UpdateGroupRoles(self)
-	wipe(roleCache)
-
-	if not self.__owner then
-		self.__owner = self:GetParent():GetParent()
-	end
-
-	local count = 0
-	for i = 1, 5 do
-		local role, class, isLeader = GetCorrectRoleInfo(self.__owner, i)
-		local roleIndex = role and roleOrder[role]
-		if roleIndex then
-			count = count + 1
-			if not roleCache[count] then roleCache[count] = {} end
-			roleCache[count][1] = roleIndex
-			roleCache[count][2] = class
-			roleCache[count][3] = isLeader
-		end
-	end
-
-	sort(roleCache, sortRoleOrder)
-end
-
-function M:ReplaceGroupRoles(numPlayers, _, disabled)
-	UpdateGroupRoles(self)
-
-	for i = 1, 5 do
-		local icon = self.Icons[i]
-		if not icon.role then
-			if i == 1 then
-				icon:SetPoint("RIGHT", -5, -2)
-			else
-				icon:ClearAllPoints()
-				icon:SetPoint("RIGHT", self.Icons[i-1], "LEFT", 2, 0)
-			end
-			icon:SetSize(26, 26)
-			icon.ClassCircle:SetSize(26, 26)
-
-			icon.RoleIcon:ClearAllPoints()
-			icon.RoleIcon:SetSize(16, 16)
-			icon.RoleIcon:SetPoint("TOPLEFT", icon, -3, 3)
-
-			icon.leader = self:CreateTexture(nil, "OVERLAY", nil, 1)
-			icon.leader:SetSize(14, 9)
-			icon.leader:SetPoint("TOP", icon, 4, 4)
-			icon.leader:SetAtlas("groupfinder-icon-leader")
-			icon.leader:SetRotation(rad(-15))
-		end
-
-		if i <= numPlayers then
-			icon.leader:SetDesaturated(disabled)
-			icon.leader:SetAlpha(disabled and .5 or 1)
-		end
-		icon.leader:Hide()
-	end
-
-	local iconIndex = numPlayers
-	for i = 1, #roleCache do
-		local roleInfo = roleCache[i]
-		if roleInfo then
-			local icon = self.Icons[iconIndex]
-			icon.ClassCircle:SetAtlas(LFG_LIST_GROUP_DATA_ATLASES[roleInfo[2]])
-			icon.leader:SetShown(roleInfo[3])
-			iconIndex = iconIndex - 1
-		end
-	end
-end
-
-function M:ReplaceGroupRoles_OLD(numPlayers, _, disabled)
-	UpdateGroupRoles(self)
-
-	for i = 1, 5 do
-		local icon = self.Icons[i]
-		if not icon.role then
-			if i == 1 then
-				icon:SetPoint("RIGHT", -5, -2)
-			else
-				icon:ClearAllPoints()
-				icon:SetPoint("RIGHT", self.Icons[i-1], "LEFT", 2, 0)
-			end
-			icon:SetSize(26, 26)
-
-			icon.role = self:CreateTexture(nil, "OVERLAY", nil, 2)
-			icon.role:SetSize(16, 16)
-			icon.role:SetPoint("TOPLEFT", icon, -3, 3)
-
-			icon.leader = self:CreateTexture(nil, "OVERLAY", nil, 1)
-			icon.leader:SetSize(14, 14)
-			icon.leader:SetPoint("TOP", icon, 4, 7)
-			icon.leader:SetTexture("Interface\\GroupFrame\\UI-Group-LeaderIcon")
-			icon.leader:SetRotation(rad(-15))
-		end
-
-		if i > numPlayers then
-			icon.role:Hide()
-		else
-			icon.role:Show()
-			icon.role:SetDesaturated(disabled)
-			icon.role:SetAlpha(disabled and .5 or 1)
-			icon.leader:SetDesaturated(disabled)
-			icon.leader:SetAlpha(disabled and .5 or 1)
-		end
-		icon.leader:Hide()
-	end
-
-	local iconIndex = numPlayers
-	for i = 1, #roleCache do
-		local roleInfo = roleCache[i]
-		if roleInfo then
-			local icon = self.Icons[iconIndex]
-			icon:SetAtlas(LFG_LIST_GROUP_DATA_ATLASES[roleInfo[2]])
-			B.ReskinSmallRole(icon.role, indexToRole[roleInfo[1]])
-			icon.leader:SetShown(roleInfo[3])
-			iconIndex = iconIndex - 1
-		end
-	end
-
-	for i = 1, iconIndex do
-		self.Icons[i].role:SetTexture(nil)
-	end
 end
 
 function M:QuickJoin_ShowTips()
@@ -327,94 +168,6 @@ function M:ReplaceFindGroupButton()
 	if C.db["Skins"]["BlizzardSkins"] then B.Reskin(bu) end
 end
 
-function M:AddDungeonsFilter()
-	local mapData = {
-		[0] = {mapID = 399, aID = 1176}, -- 红玉新生法池
-		[1] = {mapID = 400, aID = 1184}, -- 诺库德狙击战
-		[2] = {mapID = 401, aID = 1180}, -- 碧蓝魔馆
-		[3] = {mapID = 402, aID = 1160}, -- 艾杰斯亚学院
-		[4] = {mapID = 405, aID = 1164}, -- 蕨皮山谷
-		[5] = {mapID = 406, aID = 1168}, -- 注能大厅
-		[6] = {mapID = 404, aID = 1172}, -- 奈萨鲁斯
-		[7] = {mapID = 403, aID = 1188}, -- 奥丹姆：提尔的遗产
-	}
-
-	local function GetDungeonNameByID(mapID)
-		local name = C_ChallengeMode_GetMapUIInfo(mapID)
-		name = gsub(name, ".-"..HEADER_COLON, "") -- abbr Tazavesh
-		return name
-	end
-
-	local allOn
-	local filterIDs = {}
-
-	local function toggleAll()
-		allOn = not allOn
-		for i = 0, 7 do
-			mapData[i].isOn = allOn
-			filterIDs[mapData[i].aID] = allOn
-		end
-		UIDropDownMenu_Refresh(B.EasyMenu)
-		LFGListSearchPanel_DoSearch(searchPanel)
-	end
-
-	local menuList = {
-		[1] = {text = _G.SPECIFIC_DUNGEONS, isTitle = true, notCheckable = true},
-		[2] = {text = _G.SWITCH, notCheckable = true, keepShownOnClick = true, func = toggleAll},
-	}
-
-	local function onClick(self, index, aID)
-		allOn = true
-		mapData[index].isOn = not mapData[index].isOn
-		filterIDs[aID] = mapData[index].isOn
-		LFGListSearchPanel_DoSearch(searchPanel)
-	end
-
-	local function onCheck(self)
-		return mapData[self.arg1].isOn
-	end
-
-	for i = 0, 7 do
-		local value = mapData[i]
-		menuList[i+3] = {
-			text = GetDungeonNameByID(value.mapID),
-			arg1 = i,
-			arg2 = value.aID,
-			func = onClick,
-			checked = onCheck,
-			keepShownOnClick = true,
-		}
-		filterIDs[value.aID] = false
-	end
-
-	searchPanel.RefreshButton:HookScript("OnMouseDown", function(self, btn)
-		if btn ~= "RightButton" then return end
-		EasyMenu(menuList, B.EasyMenu, self, 25, 50, "MENU")
-	end)
-
-	searchPanel.RefreshButton:HookScript("OnEnter", function()
-		GameTooltip:AddLine(DB.RightButton.._G.SPECIFIC_DUNGEONS)
-		GameTooltip:Show()
-	end)
-
-	hooksecurefunc("LFGListUtil_SortSearchResults", function(results)
-		if categorySelection.selectedCategory ~= 2 then return end
-		if not allOn then return end
-
-		for i = #results, 1, -1 do
-			local resultID = results[i]
-			local searchResultInfo = C_LFGList_GetSearchResultInfo(resultID)
-			local aID = searchResultInfo and searchResultInfo.activityID
-			if aID and not filterIDs[aID] then
-				tremove(results, i)
-			end
-		end
-		searchPanel.totalResults = #results
-
-		return true
-	end)
-end
-
 local function clickSortButton(self)
 	self.__owner.Sorting.Expression:SetText(self.sortStr)
 	self.__parent.RefreshButton:Click()
@@ -513,17 +266,10 @@ function M:QuickJoin()
 
 	hooksecurefunc("StaticPopup_Show", M.HookDialogOnShow)
 	hooksecurefunc("LFGListInviteDialog_Show", M.HookDialogOnShow)
-
-	if DB.isNewPatch then
-		hooksecurefunc("LFGListGroupDataDisplayEnumerate_Update", M.ReplaceGroupRoles)
-	else
-		hooksecurefunc("LFGListGroupDataDisplayEnumerate_Update", M.ReplaceGroupRoles_OLD)
-	end
 	hooksecurefunc("LFGListSearchEntry_Update", M.ShowLeaderOverallScore)
 
 	M:AddAutoAcceptButton()
 	M:ReplaceFindGroupButton()
-	M:AddDungeonsFilter()
 	M:AddPGFSortingExpression()
 	M:FixListingTaint()
 end